Add user authentication to todo routes

+++ app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserServerSession } from "@/auth/actions/auth-actions";
 import prisma from "@/lib/prisma";
 import { NextResponse } from "next/server";
 import * as yup from "yup";
@@ -37,11 +38,26 @@ const postSchema = yup.object({
 });
 
 export async function POST(req: Request) {
+  const user = await getUserServerSession();
+  if (!user) {
+    return NextResponse.json(
+      {
+        message: "Unauthorized",
+      },
+      { status: 401 }
+    );
+  }
   try {
-    const body = await postSchema.validate(await req.json());
+    const { complete, description } = await postSchema.validate(
+      await req.json()
+    );
 
     const todo = await prisma.todo.create({
-      data: body,
+      data: {
+        complete,
+        description,
+        userId: user.id,
+      },
     });
     return NextResponse.json(todo);
   } catch (e: any) {
@@ -55,10 +71,22 @@ export async function POST(req: Request) {
 }
 
 export async function DELETE(req: Request) {
+  const user = await getUserServerSession();
+
+  if (!user) {
+    return NextResponse.json(
+      {
+        message: "Unauthorized",
+      },
+      { status: 401 }
+    );
+  }
+
   try {
     await prisma.todo.deleteMany({
       where: {
         complete: true,
+        userId: user.id,
       },
     });
     return NextResponse.json({ message: "Deleted all completed todos" });
